+++ R/Kernel_Cosine.R
@@ -39,20 +39,17 @@ Cosine$set("public","variance",function(){
   return(1 - 8/(pi^2))
 })
 Cosine$set("public","initialize",function(decorators = NULL){
-
-  pdf <- function(x1){
-    return(pi/4 * cos(pi/2 * x1))
-  }
-  cdf <- function(x1){
-    return(0.5 * (sin((pi*x1)/2)+1))
-  }
-  quantile <- function(p){
-    return((2*asin(2*p - 1))/pi)
-  }
-
-  super$initialize(decorators = decorators, pdf = pdf, cdf = cdf, quantile = quantile,
-                   support = Interval$new(-1,1),  symmetric = TRUE)
-  invisible(self)
+  super$initialize(decorators = decorators,
+                   support = Interval$new(-1,1))
+})
+Cosine$set("private",".pdf",function(x){
+  pi/4 * cos(pi/2 * x)
+})
+Cosine$set("private",".cdf",function(x){
+  0.5 * (sin((pi*x)/2)+1)
+})
+Cosine$set("private",".quantile",function(p){
+  (2*asin(2*p - 1))/pi
 })
 
 .distr6$kernels = rbind(.distr6$kernels, data.table::data.table(ShortName = "Cos", ClassName = "Cosine", Support = "[-1,1]", Packages = "-"))