+++ R/Kernel_Sigmoid.R
@@ -42,14 +42,11 @@ Sigmoid$set("public","variance",function(){
   return(pi^2/4)
 })
 Sigmoid$set("public","initialize",function(decorators = NULL){
-
-  pdf <- function(x1){
-    return((2/pi) * (exp(x1) + exp(-x1))^-1)
-  }
-
-  super$initialize(decorators = decorators, pdf = pdf, support = Reals$new(),
-                   symmetric = TRUE)
-  invisible(self)
-}) # CDF, QUANTILE & VAR MISSING
+  super$initialize(decorators = decorators,
+                   support = Reals$new())
+})
+Sigmoid$set("private",".pdf",function(x){
+  return((2/pi) * (exp(x) + exp(-x))^-1)
+})
 
 .distr6$kernels = rbind(.distr6$kernels, data.table::data.table(ShortName = "Sigm", ClassName = "Sigmoid", Support = "\u211D", Packages = "-"))