+++ tests/testthat/test-KernelTriweight.R
@@ -1,25 +1,12 @@
 library(testthat)
 
-triw = Triweight$new()
-
-test_that("zero statistics",{
-  expect_equal(triw$mean(), 0)
-  expect_equal(triw$median(), 0)
-  expect_equal(triw$mode(), 0)
+test_that("autotest",{
+  autotest_kernel(Triweight,
+                  shortname = "Triw",
+                  variance = 1/9,
+                  squared2Norm = 350/429,
+                  support = Interval$new(-1,1),
+                  pdf = c(1.0613, 1.0938, 1.0613),
+                  cdf = c(0.3917, 0.5, 0.6083))
 })
 
-test_that("var & squared-norm",{
-  expect_equal(triw$variance(), 1/9)
-  expect_equal(triw$squared2Norm(), 350/429)
-})
-
-test_that("support",{
-  expect_equal(triw$support$strprint(), Interval$new(-1,1)$strprint())
-})
-
-test_that("d/p/q/r",{
-  expect_equal(round(triw$pdf(c(-0.1,0,0.1)),4), c(1.0613, 1.0938, 1.0613))
-  expect_equal(triw$cdf(0.4), 0.873964)
-  expect_null(triw$quantile(0.42))
-  expect_null(triw$rand(10))
-})